+++ R/get_trend.R
@@ -5,7 +5,8 @@
 #' @param trend_length Number of neighbouring points to use, in days
 #' @param forecast Include forecast of component
 #' @author Daniel Ollech
-#' @examples x = daily_sim(n=4)$original # series with length 4 years
+#' @examples set.seed(123)
+#' x = daily_sim(n=4)$original # series with length 4 years
 #' res <- dsa(x, cval=7, model=c(3,1,0),fourier_number = 13, reg.create=NULL) 
 #' get_trend(res)
 #' @details If not odd the parameter trend_length is set to the next highest odd number.