+++ R/xts2ts.R
@@ -9,26 +9,48 @@
 #' @export
 
 
+
 xts2ts <- function(series, freq=NULL) {
-
-if (is.null(freq)) {freq = freq_xts(series)}  
   
-newTS <- series  
+  if (is.null(freq)) {freq = freq_xts(series)}  
   
-newTS[format(zoo::index(series), "%m-%d")=="02-29"] <- NA
-
-time <- sum(as.numeric(format(zoo::index(series), "%Y")==format(xts::first(zoo::index(series)), "%Y")))
-
-newstart <- c(as.numeric(format(xts::first(zoo::index(series)), "%Y")), (freq-time+1))
+  newTS <- series  
   
-newseries <- as.numeric(newTS[!is.na(newTS)])
-
-outTS <- stats::ts(newseries, start=newstart, frequency=freq)
-
-outTS
-
+  if (freq==365) {
+    newTS[format(zoo::index(series), "%m-%d")=="02-29"] <- NA}
+  
+  time <- sum(as.numeric(format(zoo::index(series), "%Y")==format(xts::first(zoo::index(series)), "%Y")))
+  
+  .is.leapyear <- function(x) {
+    year <- as.numeric(x)
+    if((year %% 4) == 0) {
+      if((year %% 100) == 0) {
+        if((year %% 400) == 0) {
+          out <- TRUE
+        } else {out <- FALSE}
+      } else {out <- TRUE }
+    } else {out <- FALSE}
+    
+    return(out)
+  }
+  
+  if (.is.leapyear(format(xts::first(zoo::index(series)), "%Y"))) {time <- time-1}
+  
+  newstart <- c(as.numeric(format(xts::first(zoo::index(series)), "%Y")), freq-time+1)
+  
+  newseries <- as.numeric(newTS[!is.na(newTS)])
+  
+  outTS <- stats::ts(newseries, start=newstart, frequency=freq)
+  
+  outTS
+  
 } 
-  
+
+
+
+
+
+